+++ R/context.R
@@ -29,3 +29,22 @@ context_start <- function(desc) {
     get_reporter()$.start_context(desc)
   }
 }
+
+#' Start test context from a file name
+#'
+#' For use in external reporters
+#'
+#' @param name file name
+#' @keywords internal
+#' @export
+context_start_file <- function(name) {
+  context_start(context_name(name))
+}
+
+context_name <- function(filename) {
+  # Remove test- prefix
+  filename <- sub("^test-", "", filename)
+  # Remove terminal extension
+  filename <- sub("[.][Rr]$", "", filename)
+  filename
+}